marsRover25 restructures routing and middleware files
