+++ server.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const app = express();
 const port = 3000;
-const router = require('./routes');
+const router = require('./server/routes/routes');
 const bodyParser = require('body-parser');
 
 app.use(bodyParser.json());
