+++ server/middleware/checkForUser.js
